Changed the way, how the CSS is included.
Added unix timestamp as query parameter to the css.
If there are any changes in the style.css, I'll insert a new unixtimestamp, to make sure browsers don't use the old css from their cache.
For developing, you could uncomment the script tag in head, to make sure there is always the newest style.css present in the browser.
But this is not for production environments.